Python: rename tokens module to keys
Because it's about the OpenAI API key (which is a bearer token),
but from its name it would be assumed to be abount tokens that
text is split up into for processing. (See issue #11.)

While I'm at it, I've re-run `pipenv update` to bump some minor
indirect dependencies in Pipfile.lock.

Note that the different embeddings in some of the tests are due to
neiher of these changes themselves, but to me re-running notebooks
for testing. The model is nondetermnistic (see readme) and so long
as it doesn't get in the way of the `compare/check.rb` comparison
or the comparisons in the notebook, I think it's fine to allow
these to change across commits.