Streamline Python dependencies (pipenv)
+ Re-run notebooks
+ Re-run Java code (see below for why)

The main change here is to remove the packages from Pipfile that
were only there to allow openai.embeddings_utils to be imported,
and instead depend on openai[embeddings]. The embeddings extra of
the openai package declares those packages as dependencies.

The model is nondeterministic, and getting comparable results
across separate runs is often easier when the runs are close
together in time. (I suspect this is due to caching, or to the
same or similar machines being used for requests sent close
together in time from the same IP, or a combination of the two. On
the significance of the hardware used to run a model, see
https://pytorch.org/docs/stable/notes/randomness.html.) Re-running
both the Python and Java code allows me to get results that
compare/check.rb recognized as equal (i.e., results that are equal,
up to the precision of the data types being used).