Regenerate Java embeddings again
The embeddings from Python and Java now compare equal by the Ruby
script.

As noted in some previous commit messages, because the model is
nondeterministic, it is sometimes necessary to run the Python
and/or Java code multiple times to get results that compare equal
(rather than just being very close).