Update okio
By declaring it as a direct dependency.

This is an attempt to avoid using a vulnerable version of
com.squareup.okio:okio by adding it as a direct dependency, taking
it from 3.2.0 (as resolved in the dependency graph, though 3.3.0,
which is also vulnerable, may often have been in use in practice)
to 3.4.0.

Versions prior to 3.4.0 are affected by CVE-2023-3635.