Fix JSON injection vulnerabilty in demo script
This makes the full length demo script, "shell/demo", more robust.
In contrast, it does not change the behavior of the short demo
script "shell/demo-short", which is modified only to update the
way its relationship to "shell/demo" is described (now that they
are not both subject to JSON injection, but only the short one).

Although the vulnerability was documented and clearly warned about
in the comments in "shell/demo", the point of that longer script is
to show the technique in somewhat more production-like code. So it
never really make sense for it to be that brittle. This fixes that,
by using jq to build the JSON in a way that automatically applies
proper escaping sufficient even for adversarial input.

This is still not fully production quality, nor is it intended to
be. Clarity of presentation remains the primary goal. The request
is assumed to succeed, and short but uninformative output still
results if the request fails due to an absent or invalid API key or
for any other reason.